Complex source

+++ src/cogradio_utils/reconstruction/crosscorr.py
@@ -31,7 +31,7 @@ class CrossCorrelation(Reconstructor):
         return rx
 
     def cross_correlation_signals(self, signal):
-        Ry = np.zeros((self.M ** 2, 2 * self.L - 1))
+        Ry = np.zeros((self.M ** 2, 2 * self.L - 1), dtype=np.complex64)
         for i in range(self.M):
             for j in range(self.M):
                 Ry[i * self.M + j] = np.correlate(signal[i, :],
